requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.12
distlib==0.3.4
Django==3.2.6
django-cors-headers==3.11.0
djangorestframework==3.13.1
filelock==3.6.0
idna==3.3
mysqlclient==2.1.0
platformdirs==2.5.1
pytz==2021.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.8
virtualenv==20.13.2
